# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the BuffetPOS API documentation.",
        "title": "BuffetPOS API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/customer/orders": {
            "post": {
                "description": "Add order to table.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Add Order",
                "parameters": [
                    {
                        "description": "User Add Order Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserAddOrderRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access Code",
                        "name": "AccessCode",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/customer/orders/history": {
            "get": {
                "description": "Get order history for table.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Get Order History",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Code",
                        "name": "AccessCode",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.OrderDetail"
                            }
                        }
                    }
                }
            }
        },
...
        "/manage/menus/best-selling": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Best Selling Menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Get Best Selling Menu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.NumberMenu"
                            }
                        }
                    }
                }
            }
        },
...
        "/manage/orders": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get orders by status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Get Orders By Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status of order",
                        "name": "status",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.OrderDetail"
                            }
                        }
                    }
                }
            }
        },
        "/manage/orders/status": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update order status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Update Order Status",
                "parameters": [
                    {
                        "description": "Update Order Status Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateOrderStatusRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/manage/orders/tables": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get orders by table ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Get Orders By Table ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Table ID",
                        "name": "tableID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.OrderDetail"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.OrderItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "menuID": {
                    "type": "string"
                },
                "orderID": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "models.OrderStatus": {
            "type": "string",
            "enum": [
                "preparing",
                "served",
                "succeeded",
                "cancelled"
            ],
            "x-enum-varnames": [
                "PREPARING",
                "SERVED",
                "SUCCEEDED",
                "CANCELLED"
            ]
        },
...
        "requests.OrderItemRequest": {
            "type": "object",
            "required": [
                "menu_id",
                "quantity"
            ],
            "properties": {
                "menu_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "requests.UpdateOrderStatusRequest": {
            "type": "object",
            "required": [
                "status",
                "table_id"
            ],
            "properties": {
                "status": {
                    "type": "string"
                },
                "table_id": {
                    "type": "string"
                }
            }
        },
        "requests.UserAddOrderRequest": {
            "type": "object",
            "required": [
                "order_items"
            ],
            "properties": {
                "order_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.OrderItemRequest"
                    }
                }
            }
        },
...
        "responses.NumberMenu": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isAvailable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                }
            }
        },
        "responses.OrderDetail": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "orderItem": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItem"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.OrderStatus"
                },
                "tableId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}